Refactor TOMLFile logic
This change allows for peotry to have a shared interface to tomlkit's
TOMLFile class.
