+++ poetry/core/pyproject/__init__.py
@@ -1,10 +1,6 @@
 from poetry.core.pyproject.exceptions import PyProjectException
 from poetry.core.pyproject.tables import BuildSystem
 from poetry.core.pyproject.toml import PyProjectTOML
-from poetry.core.pyproject.toml import PyProjectTOMLFile
 
 
-__all__ = [
-    clazz.__name__
-    for clazz in {BuildSystem, PyProjectException, PyProjectTOML, PyProjectTOMLFile}
-]
+__all__ = [clazz.__name__ for clazz in {BuildSystem, PyProjectException, PyProjectTOML}]
