+++ poetry/core/toml/exceptions.py
@@ -1,0 +1,7 @@
+from tomlkit.exceptions import TOMLKitError
+
+from poetry.core.exceptions import PoetryCoreException
+
+
+class TOMLError(TOMLKitError, PoetryCoreException):
+    pass
