+++ poetry/core/utils/toml_file.py
@@ -1,18 +1,15 @@
 # -*- coding: utf-8 -*-
-from poetry.core.pyproject import PyProjectTOMLFile
+from poetry.core.toml import TOMLFile
 
 
-class TomlFile(PyProjectTOMLFile):
+class TomlFile(TOMLFile):
     @classmethod
     def __new__(cls, *args, **kwargs):
         import warnings
 
         warnings.warn(
             "Use of {}.{} has been deprecated, use {}.{} instead.".format(
-                cls.__module__,
-                cls.__name__,
-                PyProjectTOMLFile.__module__,
-                PyProjectTOMLFile.__name__,
+                cls.__module__, cls.__name__, TOMLFile.__module__, TOMLFile.__name__,
             ),
             category=DeprecationWarning,
             stacklevel=2,
